+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-# 0.3.1
+# 0.4.0
 
+-   Added onKeyPress / onFocus / onBlur callbacks for all components with the option to return false to prevent the default action
+-   Fixed issue with addChild if the child has already been added
+-   Added position type for relative/absolute positioning
+-   Added support for percentage based x/y
 -   Updated documentation and added logger options to the main initialize fuunction
 
 # 0.3.0